Move used-once helpers out of Address model methods.
